Reduces nonvirtual block count of the GUAM model below 1000 so as to be compatible with Home and Student Matlab licenses.
This is done mainly by replacing certain library blocks which were previously implemented with Simulink blocks with Matlab functions.
These are found in the starslibAxisTransformations, starslibKinematicalEquations, and starslibMathOperations libraries; the new blocks have furthermore been propagated to the (otherwise identical) libraries libAxisTransformations and libMathOperations via library links.
In all cases, the former implementations have been saved in blocks with the prefix 'slk' in the library block name, and the new implementation with prefix 'fun'.
The 'active' block then keeps the old name to preserve library links, and is simply a library link to the 'fun' or 'slk' implementation.

The 'Qmult' and 'Qtrans' functions, which are called from the embedded Matlab functions, have been slightly modified to allow for code generation.

NB: modified files are compatible with Matlab release R2024b; a future update will restore them to an earlier version.